rearrange files, cmake install, texture limits
files got rearranged a bit, cmake install exists (idk), limit on texture size, limit on number of loadable buffers and textures.
