+++ examples/verbosebuffer.cpp
