+++ examples/verbosetexture.cpp
